ENH: moved run monitoring code to a separate file
